+++ Branch-SDK/Branch-SDK/BNCDeviceInfo.h
@@ -35,5 +35,7 @@
 
 //----------Methods----------------//
 + (BNCDeviceInfo *)getInstance;
++ (NSString*) userAgentString;          // Warning:  Has an implied lock on main thread on first call.
++ (NSString*) systemBuildVersion;
 
 @end