+++ Branch-SDK/Branch-SDK/BNCPreferenceHelper.m
@@ -47,7 +47,10 @@ NSString * const BRANCH_PREFS_KEY_ANALYTICS_MANIFEST = @"bnc_branch_analytics_ma
 // The name of this key was specified in the account-creation API integration
 static NSString * const BNC_BRANCH_FABRIC_APP_KEY_KEY = @"branch_key";
 
-@interface BNCPreferenceHelper ()
+@interface BNCPreferenceHelper () {
+    NSString *_lastSystemBuildVersion;
+    NSString *_browserUserAgentString;
+}
 
 @property (strong, nonatomic) NSMutableDictionary *persistenceDict;
 @property (strong, nonatomic) NSMutableDictionary *creditsDictionary;
@@ -163,6 +166,8 @@ static NSString * const BNC_BRANCH_FABRIC_APP_KEY_KEY = @"branch_key";
     NSUInteger index = BNC_API_BASE_URL.length;
     return [url substringFromIndex:index];
 }
+
+
 #pragma mark - Preference Storage
 
 - (NSString *)getBranchKey:(BOOL)isLive {
@@ -432,6 +437,34 @@ static NSString * const BNC_BRANCH_FABRIC_APP_KEY_KEY = @"branch_key";
         _appleSearchAdDetails = (NSDictionary *) [self readObjectFromDefaults:BRANCH_PREFS_KEY_APPLE_SEARCH_ADS_INFO];
     }
     return [_appleSearchAdDetails isKindOfClass:[NSDictionary class]] ? _appleSearchAdDetails : nil;
+}
+
+- (NSString*) lastSystemBuildVersion {
+    if (!_lastSystemBuildVersion) {
+        _lastSystemBuildVersion = [self readStringFromDefaults:@"_lastSystemBuildVersion"];
+    }
+    return _lastSystemBuildVersion;
+}
+
+- (void) setLastSystemBuildVersion:(NSString *)lastSystemBuildVersion {
+    if (![_lastSystemBuildVersion isEqualToString:lastSystemBuildVersion]) {
+        _lastSystemBuildVersion = lastSystemBuildVersion;
+        [self writeObjectToDefaults:@"_lastSystemBuildVersion" value:_lastSystemBuildVersion];
+    }
+}
+
+- (NSString*) browserUserAgentString {
+    if (!_browserUserAgentString) {
+        _browserUserAgentString = [self readStringFromDefaults:@"_browserUserAgentString"];
+    }
+    return _browserUserAgentString;
+}
+
+- (void) setBrowserUserAgentString:(NSString *)browserUserAgentString {
+    if (![_browserUserAgentString isEqualToString:browserUserAgentString]) {
+        _browserUserAgentString = browserUserAgentString;
+        [self writeObjectToDefaults:@"_browserUserAgentString" value:_browserUserAgentString];
+    }
 }
 
 - (NSString *)userUrl {